+++ frog-utils/src/ReactiveRichText/constants.js
@@ -1,0 +1,25 @@
+const LiViewTypes = {
+  VIEW: 'view',
+  THUMB: 'thumbView',
+  EDIT: 'edit'
+};
+
+const formats = [
+  'bold',
+  'italic',
+  'underline',
+  'strike',
+  'blockquote',
+  'code-block',
+  'header',
+  'list',
+  'link',
+  'image',
+  'video',
+  'learning-item',
+  'author',
+  'li-view',
+  'background'
+];
+
+export { LiViewTypes, formats };
